added cache use cases

+++ app/src/main/java/com/example/treedatabase/data/model/NodeData.kt
@@ -1,3 +1,3 @@
 package com.example.treedatabase.data.model
 
-data class NodeData(val id: Int, val value: String, val parentId: Int?, val deleted: Boolean)
+data class NodeData(val id: Long, val value: String, val parentId: Long?, val deleted: Boolean)
